AAAaaargh ugly default.conf added.
Open tree web apps run as an unprivileged user, without
a home dir, a bare bones env, and are installed as an
unprivileged user.
So it is hard to use any location for the peyotl.conf file
(other than /tmp).

This commit creates a peyotl/default.conf file that is
used if ~/.peyotl/config is not found. The defaults
in that file are subject to change.

Users should make use of their ~/.peyotl/config file
to be resistant to changes in the file.

Currently it produced verbose logs at /tmp/peyotl-log