Making tests work on naive machine. Minor commit.
This commit skips tests that are only work after
some special configuration. These tests are only regularly
run by MTH and EJBM (and they should still run
with these checks in place).
This is a minor commit, but it is a bug-fix in the
sense that it cleans up troubling errors if you
try to run the tests on a naive machine.