Smarter History tab (show target file only!)
The reported version history of a study was occasionally showing other
study Nexson files, instead of just the one we're interested in, most
obviously in the History tab of our study curation app.

This was due to the '--follow' option in 'git log', which assumes that
very similar files are in fact the same. Smart for source code, but dumb
for Nexson files, which can be 99% the same but distinct.