Discovered a --py3k option to pylint...
This will warn about constructs in py2 that are not supported
in py3, and should cut down on the # of times I commit something,
push it, and then have to fix it when travis tells me that the
py3 version is broken.
the from __future__ imports change the behavior of print, the
division operator (for ints) and require the use of absolute
imports (which is a convention that we tried to follow anyway,
but which will now be enforced).