Using a post-merge hook on a commit to reg. ID
If a study is deleted and then reinstated, we want its
ID to go back in the indices that hold study ID info.
These indices refer to the master branch only, so they have to
be updated when a study is created, deleted (if it merges), or
reinstated (committed and merged to master, but not a new
study ID).
These indices aren't used in many calls, but we should keep them
up-to-date, and this commit does that.