Allowing non-phylesystem shards in the parent.
This bug was reported by Karen Cranston.
If you had a git repo that was not a phylesystem shard in
the parent of your phylesystem, then the Phylesystem
wrapper should just skip over non-ps dirs. but it was
failing with an exception.

This commit catches and warns rather than crashing.