Fix for broken test; move to an earlier commit SHA
`testChangedStudies` doesn't match the history of this test repo;
we're looking for studies that have changed since the most recent
commit, which will always be empty. The previous SHA fits the bill, and
substituting that satisfies all tests.

Here's the change where this SHA was entered:
https://github.com/OpenTreeOfLife/peyotl/commit/cd157566dda7297e929bbe80e0afc6c5fca55adc

... and here's the test repo, where
2d59ab892ddb3d09d4b18c91470b8c1c4cca86dc is still the most recent
commit:
https://github.com/mtholder/mini_phyl/commits/master

Perhaps this test would have passed against the original (snacktavish)
test repo?
https://github.com/snacktavish/mini_phyl/commits/master

@mtholder, please review these links and feel free to revert this if
I've misunderstood this test.