Bug fixes in pruning forwarded IDs.
Wrong var name for ott_id and failure to set value for key.
The latter would have crashed if it had ever been triggered, because
the dict is a property (so can't be set).
The former would have crashed because it tried to double delete
the ott_id key set in a previous loop (and already deleted there).